Extract Modal Peaks

Extracts modal peak information from each sample in a list of fragments.

Usage

extract_alleles(fragments_list)

Arguments

fragments_list

A list of fragments objects containing fragment data.

Value

A dataframe containing modal peak information for each sample

Examples

fsa_list <- lapply(cell_line_fsa_list, function(x) x$clone())
# import data with read_fsa() to generate an equivalent list to cell_line_fsa_list
test_fragments <- trace(fsa_list, grouped = TRUE, metadata_data.frame = metadata)
#> Finding ladders
#>   |                                                                              |                                                                      |   0%  |                                                                              |====                                                                  |   5%  |                                                                              |=======                                                               |  11%  |                                                                              |===========                                                           |  16%  |                                                                              |===============                                                       |  21%  |                                                                              |==================                                                    |  26%  |                                                                              |======================                                                |  32%  |                                                                              |==========================                                            |  37%  |                                                                              |=============================                                         |  42%  |                                                                              |=================================                                     |  47%  |                                                                              |=====================================                                 |  53%  |                                                                              |=========================================                             |  58%  |                                                                              |============================================                          |  63%  |                                                                              |================================================                      |  68%  |                                                                              |====================================================                  |  74%  |                                                                              |=======================================================               |  79%  |                                                                              |===========================================================           |  84%  |                                                                              |===============================================================       |  89%  |                                                                              |==================================================================    |  95%  |                                                                              |======================================================================| 100%
#> Finding fragments
#>   |                                                                              |                                                                      |   0%  |                                                                              |====                                                                  |   5%  |                                                                              |=======                                                               |  11%  |                                                                              |===========                                                           |  16%  |                                                                              |===============                                                       |  21%  |                                                                              |==================                                                    |  26%  |                                                                              |======================                                                |  32%  |                                                                              |==========================                                            |  37%  |                                                                              |=============================                                         |  42%  |                                                                              |=================================                                     |  47%  |                                                                              |=====================================                                 |  53%  |                                                                              |=========================================                             |  58%  |                                                                              |============================================                          |  63%  |                                                                              |================================================                      |  68%  |                                                                              |====================================================                  |  74%  |                                                                              |=======================================================               |  79%  |                                                                              |===========================================================           |  84%  |                                                                              |===============================================================       |  89%  |                                                                              |==================================================================    |  95%  |                                                                              |======================================================================| 100%
#> Finding alleles
#> Calling repeats
#> Assigning index peaks

extract_alleles(test_fragments)
#>                                   unique_id allele_size allele_signal
#> 20230413_A07.fsa           20230413_A07.fsa    481.5436          1309
#> 20230413_A08.fsa           20230413_A08.fsa    481.5379          1307
#> 20230413_A09.fsa           20230413_A09.fsa    487.1819          1361
#> 20230413_C01.fsa           20230413_C01.fsa    483.4888          1161
#> 20230413_C02.fsa           20230413_C02.fsa    481.1637          1248
#> 20230413_C03.fsa           20230413_C03.fsa    481.3875          1157
#> 20230413_D07.fsa           20230413_D07.fsa    479.0952          1456
#> 20230413_D08.fsa           20230413_D08.fsa    476.3517          1465
#> 20230413_D09.fsa           20230413_D09.fsa    476.4824          1189
#> 20230413_F01.fsa           20230413_F01.fsa    477.6280           862
#> 20230413_F02.fsa           20230413_F02.fsa    475.3879          1056
#> 20230413_F03.fsa           20230413_F03.fsa    478.4716          1013
#> 20230413_G07.fsa           20230413_G07.fsa    467.2601          3585
#> 20230413_G08.fsa           20230413_G08.fsa    467.5702          2025
#> 20230413_G09.fsa           20230413_G09.fsa    467.3300          4201
#> 20230413_H07.fsa           20230413_H07.fsa    415.3404          3948
#> 20230413_H08.fsa           20230413_H08.fsa    421.3037          2152
#> S-21-211_20220630.fsa S-21-211_20220630.fsa    413.6654          2025
#> S-21-212_20220630.fsa S-21-212_20220630.fsa    419.4859          1996
#>                       allele_repeat allele_2_size allele_2_signal
#> 20230413_A07.fsa           131.5145            NA              NA
#> 20230413_A08.fsa           131.5126            NA              NA
#> 20230413_A09.fsa           133.3940            NA              NA
#> 20230413_C01.fsa           132.1629            NA              NA
#> 20230413_C02.fsa           131.3879            NA              NA
#> 20230413_C03.fsa           131.4625            NA              NA
#> 20230413_D07.fsa           130.6984            NA              NA
#> 20230413_D08.fsa           129.7839            NA              NA
#> 20230413_D09.fsa           129.8275            NA              NA
#> 20230413_F01.fsa           130.2093            NA              NA
#> 20230413_F02.fsa           129.4626            NA              NA
#> 20230413_F03.fsa           130.4905            NA              NA
#> 20230413_G07.fsa           126.7534            NA              NA
#> 20230413_G08.fsa           126.8567            NA              NA
#> 20230413_G09.fsa           126.7767            NA              NA
#> 20230413_H07.fsa           109.4468            NA              NA
#> 20230413_H08.fsa           111.4346            NA              NA
#> S-21-211_20220630.fsa      108.8885            NA              NA
#> S-21-212_20220630.fsa      110.8286            NA              NA
#>                       allele_2_repeat
#> 20230413_A07.fsa                   NA
#> 20230413_A08.fsa                   NA
#> 20230413_A09.fsa                   NA
#> 20230413_C01.fsa                   NA
#> 20230413_C02.fsa                   NA
#> 20230413_C03.fsa                   NA
#> 20230413_D07.fsa                   NA
#> 20230413_D08.fsa                   NA
#> 20230413_D09.fsa                   NA
#> 20230413_F01.fsa                   NA
#> 20230413_F02.fsa                   NA
#> 20230413_F03.fsa                   NA
#> 20230413_G07.fsa                   NA
#> 20230413_G08.fsa                   NA
#> 20230413_G09.fsa                   NA
#> 20230413_H07.fsa                   NA
#> 20230413_H08.fsa                   NA
#> S-21-211_20220630.fsa              NA
#> S-21-212_20220630.fsa              NA