Extract All Fragments

Extracts peak data from each sample in a list of fragments.

Usage

extract_fragments(fragments_list)

Arguments

fragments_list

A list of fragments_repeats objects containing fragment data.

Value

A dataframe containing peak data for each sample

Examples

gm_raw <- trace::example_data
metadata <- trace::metadata

test_fragments <- peak_table_to_fragments(gm_raw,
  data_format = "genemapper5",
  dye_channel = "B",
  min_size_bp = 400
)
#> Warning: Size filtering removed all rows for 20230413_D11.fsa

add_metadata(
  fragments_list = test_fragments,
  metadata_data.frame = metadata
)

find_alleles(
  fragments_list = test_fragments
)

call_repeats(
  fragments_list = test_fragments,
  repeat_calling_algorithm = "simple",
  assay_size_without_repeat = 87,
  repeat_size = 3
)
#> Warning: Repeats were not called in the following samples (no allele in sample): 20230413_A04.fsa, 20230413_D11.fsa

extract_alleles(test_fragments)
#>                                   unique_id   size  repeats height
#> 20230413_A01.fsa           20230413_A01.fsa 465.82 126.2733   1051
#> 20230413_A02.fsa           20230413_A02.fsa 463.36 125.4533   1096
#> 20230413_A03.fsa           20230413_A03.fsa 466.15 126.3833   1210
#> 20230413_A04.fsa           20230413_A04.fsa     NA       NA     NA
#> 20230413_A05.fsa           20230413_A05.fsa 485.98 132.9933    797
#> 20230413_A06.fsa           20230413_A06.fsa 502.08 138.3600    959
#> 20230413_A07.fsa           20230413_A07.fsa 480.54 131.1800   1240
#> 20230413_A08.fsa           20230413_A08.fsa 480.54 131.1800   1230
#> 20230413_A09.fsa           20230413_A09.fsa 486.32 133.1067   1284
#> 20230413_A10.fsa           20230413_A10.fsa 505.21 139.4033   1594
#> 20230413_A11.fsa           20230413_A11.fsa 504.69 139.2300   1554
#> 20230413_A12.fsa           20230413_A12.fsa 504.29 139.0967   1272
#> 20230413_B01.fsa           20230413_B01.fsa 563.96 158.9867    537
#> 20230413_B02.fsa           20230413_B02.fsa 560.83 157.9433   1031
#> 20230413_B03.fsa           20230413_B03.fsa 561.04 158.0133    695
#> 20230413_B04.fsa           20230413_B04.fsa 549.17 154.0567    871
#> 20230413_B05.fsa           20230413_B05.fsa 548.96 153.9867    913
#> 20230413_B06.fsa           20230413_B06.fsa 543.13 152.0433    722
#> 20230413_B07.fsa           20230413_B07.fsa 472.28 128.4267     38
#> 20230413_B08.fsa           20230413_B08.fsa 461.33 124.7767    665
#> 20230413_B09.fsa           20230413_B09.fsa 463.92 125.6400    762
#> 20230413_B10.fsa           20230413_B10.fsa 492.43 135.1433    961
#> 20230413_B11.fsa           20230413_B11.fsa 491.98 134.9933    879
#> 20230413_B12.fsa           20230413_B12.fsa 482.52 131.8400   1355
#> 20230413_C01.fsa           20230413_C01.fsa 482.43 131.8100   1068
#> 20230413_C02.fsa           20230413_C02.fsa 480.10 131.0333   1164
#> 20230413_C03.fsa           20230413_C03.fsa 480.30 131.1000   1083
#> 20230413_C04.fsa           20230413_C04.fsa 508.75 140.5833   1094
#> 20230413_C05.fsa           20230413_C05.fsa 508.54 140.5133   1123
#> 20230413_C06.fsa           20230413_C06.fsa 509.15 140.7167   1402
#> 20230413_C07.fsa           20230413_C07.fsa 569.58 160.8600    718
#> 20230413_C08.fsa           20230413_C08.fsa 569.80 160.9333    836
#> 20230413_C09.fsa           20230413_C09.fsa 567.50 160.1667    790
#> 20230413_C10.fsa           20230413_C10.fsa 548.98 153.9933    835
#> 20230413_C11.fsa           20230413_C11.fsa 546.12 153.0400    751
#> 20230413_C12.fsa           20230413_C12.fsa 552.04 155.0133    845
#> 20230413_D01.fsa           20230413_D01.fsa 462.90 125.3000   1466
#> 20230413_D02.fsa           20230413_D02.fsa 460.65 124.5500   1375
#> 20230413_D03.fsa           20230413_D03.fsa 460.94 124.6467   1562
#> 20230413_D04.fsa           20230413_D04.fsa 483.21 132.0700    676
#> 20230413_D05.fsa           20230413_D05.fsa 480.48 131.1600    469
#> 20230413_D06.fsa           20230413_D06.fsa 486.52 133.1733    457
#> 20230413_D07.fsa           20230413_D07.fsa 478.16 130.3867   1361
#> 20230413_D08.fsa           20230413_D08.fsa 475.33 129.4433   1382
#> 20230413_D09.fsa           20230413_D09.fsa 475.37 129.4567   1123
#> 20230413_D10.fsa           20230413_D10.fsa 512.29 141.7633    692
#> 20230413_D11.fsa           20230413_D11.fsa     NA       NA     NA
#> 20230413_D12.fsa           20230413_D12.fsa 508.37 140.4567   1230
#> 20230413_E01.fsa           20230413_E01.fsa 563.54 158.8467    511
#> 20230413_E02.fsa           20230413_E02.fsa 562.65 158.5500    509
#> 20230413_E03.fsa           20230413_E03.fsa 567.02 160.0067    595
#> 20230413_E04.fsa           20230413_E04.fsa 549.36 154.1200    396
#> 20230413_E05.fsa           20230413_E05.fsa 543.62 152.2067    489
#> 20230413_E06.fsa           20230413_E06.fsa 546.38 153.1267    515
#> 20230413_E07.fsa           20230413_E07.fsa 458.57 123.8567   1254
#> 20230413_E08.fsa           20230413_E08.fsa 461.43 124.8100   1618
#> 20230413_E09.fsa           20230413_E09.fsa 464.23 125.7433   1494
#> 20230413_E10.fsa           20230413_E10.fsa 478.00 130.3333    583
#> 20230413_E11.fsa           20230413_E11.fsa 477.82 130.2733    823
#> 20230413_E12.fsa           20230413_E12.fsa 480.13 131.0433    989
#> 20230413_F01.fsa           20230413_F01.fsa 476.53 129.8433    807
#> 20230413_F02.fsa           20230413_F02.fsa 474.28 129.0933   1001
#> 20230413_F03.fsa           20230413_F03.fsa 477.33 130.1100    948
#> 20230413_F04.fsa           20230413_F04.fsa 498.71 137.2367   1075
#> 20230413_F05.fsa           20230413_F05.fsa 502.13 138.3767    643
#> 20230413_F06.fsa           20230413_F06.fsa 505.53 139.5100    284
#> 20230413_F07.fsa           20230413_F07.fsa 564.89 159.2967    651
#> 20230413_F08.fsa           20230413_F08.fsa 564.89 159.2967    366
#> 20230413_F09.fsa           20230413_F09.fsa 573.83 162.2767    256
#> 20230413_F10.fsa           20230413_F10.fsa 549.17 154.0567    536
#> 20230413_F11.fsa           20230413_F11.fsa 542.86 151.9533    685
#> 20230413_F12.fsa           20230413_F12.fsa 542.65 151.8833    572
#> 20230413_G01.fsa           20230413_G01.fsa 451.96 121.6533   1101
#> 20230413_G02.fsa           20230413_G02.fsa 452.61 121.8700   2877
#> 20230413_G03.fsa           20230413_G03.fsa 452.82 121.9400    258
#> 20230413_G04.fsa           20230413_G04.fsa 471.85 128.2833   3452
#> 20230413_G05.fsa           20230413_G05.fsa 471.58 128.1933   2988
#> 20230413_G06.fsa           20230413_G06.fsa 471.89 128.2967   2844
#> 20230413_G07.fsa           20230413_G07.fsa 466.33 126.4433   3363
#> 20230413_G08.fsa           20230413_G08.fsa 466.67 126.5567   1918
#> 20230413_G09.fsa           20230413_G09.fsa 466.32 126.4400   3935
#> 20230413_G10.fsa           20230413_G10.fsa 498.92 137.3067   2957
#> 20230413_G11.fsa           20230413_G11.fsa 498.71 137.2367    154
#> 20230413_G12.fsa           20230413_G12.fsa 495.11 136.0367   2950
#> 20230413_H01.fsa           20230413_H01.fsa 545.31 152.7700   2473
#> 20230413_H02.fsa           20230413_H02.fsa 545.51 152.8367   1973
#> 20230413_H03.fsa           20230413_H03.fsa 543.00 152.0000   1896
#> 20230413_H04.fsa           20230413_H04.fsa 524.47 145.8233   2013
#> 20230413_H05.fsa           20230413_H05.fsa 526.88 146.6267   1741
#> 20230413_H06.fsa           20230413_H06.fsa 524.38 145.7933   1340
#> 20230413_H07.fsa           20230413_H07.fsa 414.55 109.1833   3798
#> 20230413_H08.fsa           20230413_H08.fsa 420.70 111.2333   2081
#> S-21-211_20220630.fsa S-21-211_20220630.fsa 414.83 109.2767   3159
#> S-21-212_20220630.fsa S-21-212_20220630.fsa 420.87 111.2900   2653