Extract ladder summary

Extract a table summarizing the ladder models

Usage

extract_ladder_summary(fragments_trace_list, sort = FALSE)

Arguments

fragments_trace_list

a list of fragments trace objects

sort

A logical statement for if the samples should be ordered by average ladder R-squared.

Value

a dataframe of ladder quality information

Details

The ladder peaks are assigned using a custom algorithm that maximizes the fit of detected ladder peaks and given base-pair sizes. This function summarizes the R-squared values of these individual correlations.

Examples

  fsa_list <- lapply(cell_line_fsa_list, function(x) x$clone())

  find_ladders(fsa_list, show_progress_bar = FALSE)

  extract_ladder_summary(fsa_list, sort = TRUE)
#>                                   unique_id   avg_rsq   min_rsq
#> S-21-211_20220630.fsa S-21-211_20220630.fsa 0.9995315 0.9987599
#> S-21-212_20220630.fsa S-21-212_20220630.fsa 0.9995515 0.9988479
#> 20230413_F01.fsa           20230413_F01.fsa 0.9995721 0.9989598
#> 20230413_H07.fsa           20230413_H07.fsa 0.9995995 0.9988479
#> 20230413_F02.fsa           20230413_F02.fsa 0.9996156 0.9989598
#> 20230413_C01.fsa           20230413_C01.fsa 0.9996193 0.9988872
#> 20230413_A08.fsa           20230413_A08.fsa 0.9996200 0.9986115
#> 20230413_G09.fsa           20230413_G09.fsa 0.9996218 0.9989546
#> 20230413_G08.fsa           20230413_G08.fsa 0.9996224 0.9986115
#> 20230413_C02.fsa           20230413_C02.fsa 0.9996303 0.9989546
#> 20230413_D07.fsa           20230413_D07.fsa 0.9996312 0.9986115
#> 20230413_H08.fsa           20230413_H08.fsa 0.9996397 0.9989546
#> 20230413_G07.fsa           20230413_G07.fsa 0.9996454 0.9989546
#> 20230413_A09.fsa           20230413_A09.fsa 0.9996507 0.9989546
#> 20230413_F03.fsa           20230413_F03.fsa 0.9996514 0.9989261
#> 20230413_D08.fsa           20230413_D08.fsa 0.9996529 0.9986115
#> 20230413_C03.fsa           20230413_C03.fsa 0.9996544 0.9989261
#> 20230413_A07.fsa           20230413_A07.fsa 0.9996898 0.9990775
#> 20230413_D09.fsa           20230413_D09.fsa 0.9997005 0.9990775