fragments_trace object

An R6 Class representing a fragments_trace object.

Details

The idea behind this class is to store data for processing of the continuous trace-level information from an fsa file towards peak level data.

It also contains methods for plotting the ladder and traces

Super class

trace::fragments -> fragments_trace

Public fields

unique_id

unique id of the sample usually the file name

fsa

The whole fsa file, output from seqinr::read.abif()

raw_ladder

The raw data from the ladder channel

raw_data

The raw data from the sample channel

scan

The scan number

off_scale_scans

vector indicating which scales were too big and off scale. Note can be in any channel

ladder_df

A dataframe of the identified ladder from find_ladders(). Scan is the scan number of peak and size is the associated bp size.

trace_bp_df

A dataframe of bp size for every scan from find_ladders().

Methods

Public methods

fragments_trace$new()

fragments_trace$plot_ladder()

fragments_trace$plot_data_channels()

fragments_trace$clone()

Inherited methods

trace::fragments$plot_trace()

trace::fragments$print()

Method new()

Create a new fragments_trace.

Usage

fragments_trace$new(unique_id, fsa_file)

Arguments

unique_id

usually the file name

fsa_file

output from seqinr::read.abif()

Returns

A new fragments_trace object.

Method plot_ladder()

plot the ladder data

Usage

fragments_trace$plot_ladder(xlim = NULL, ylim = NULL, plot_title = NULL)

Arguments

xlim

numeric vector length two specifying the x axis limits

ylim

numeric vector length two specifying the y axis limits

plot_title

A character string for setting the plot title. Defaults to the unique id of the object

Returns

A base R plot

Method plot_data_channels()

plot the raw data channels in the fsa file. It identifies every channel that has "DATA" in its name.

Usage

fragments_trace$plot_data_channels()

Returns

A base R plot

Method clone()

The objects of this class are cloneable with this method.

Usage

fragments_trace$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.