Convert Peak Table to Fragments_repeats class

This function converts a peak table data frame into a list of fragments_repeats objects.

Usage

peak_table_to_fragments(
  df,
  data_format = NULL,
  peak_size_col = NULL,
  peak_signal_col = NULL,
  unique_id = NULL,
  dye_col = NULL,
  dye_channel = NULL,
  allele_col = NULL,
  min_size_bp = 200,
  max_size_bp = 1000
)

Arguments

df

A data frame containing the peak data.

data_format

The format that the data frame is in (for example, a genemapper peak table). Choose between: genemapper5, generic.

peak_size_col

A character string specifying column name giving the peak size.

peak_signal_col

A character string specifying column name giving the peak signal.

unique_id

A character string specifying column name giving the unique sample id (often the file name).

dye_col

Genemapper specific. A character string specifying column name indicating the dye channel.

dye_channel

Genemapper specific. A character string indicating the channel to extract data from. For example, 6-FAM is often "B".

allele_col

Genemapper specific. A character string specifying column name indicating the called alleles. This is often used when the peaks have been called in genemapper.

min_size_bp

Numeric value indicating the minimum size of the peak table to import.

max_size_bp

Numeric value indicating the maximum size of the peak table to import.

Value

A list of fragments_repeats objects.

Details

This function takes a peak table data frame (eg. Genemapper output) and converts it into a list of fragment objects. The function supports different data formats and allows specifying column names for various attributes.

See also

Examples

gm_raw <- trace::example_data

test_fragments <- peak_table_to_fragments(
  gm_raw,
  data_format = "genemapper5",
  dye_channel = "B",
  min_size_bp = 400
)