Extract Modal Peaks

Extracts modal peak information from each sample in a list of fragments.

Usage

extract_alleles(fragments_list)

Arguments

fragments_list

A list of fragments_repeats objects containing fragment data.

Value

A dataframe containing modal peak information for each sample

Examples

gm_raw <- trace::example_data

test_fragments <- peak_table_to_fragments(gm_raw,
  data_format = "genemapper5",
  dye_channel = "B",
  min_size_bp = 400
)
#> Warning: Size filtering removed all rows for 20230413_D11.fsa

find_alleles(
  fragments_list = test_fragments,
  peak_region_size_gap_threshold = 6,
  peak_region_height_threshold_multiplier = 1
)

extract_alleles(test_fragments)
#>                                   unique_id   size repeats height
#> 20230413_A01.fsa           20230413_A01.fsa 465.82      NA   1051
#> 20230413_A02.fsa           20230413_A02.fsa 463.36      NA   1096
#> 20230413_A03.fsa           20230413_A03.fsa 466.15      NA   1210
#> 20230413_A04.fsa           20230413_A04.fsa     NA      NA     NA
#> 20230413_A05.fsa           20230413_A05.fsa 485.98      NA    797
#> 20230413_A06.fsa           20230413_A06.fsa 502.08      NA    959
#> 20230413_A07.fsa           20230413_A07.fsa 480.54      NA   1240
#> 20230413_A08.fsa           20230413_A08.fsa 480.54      NA   1230
#> 20230413_A09.fsa           20230413_A09.fsa 486.32      NA   1284
#> 20230413_A10.fsa           20230413_A10.fsa 505.21      NA   1594
#> 20230413_A11.fsa           20230413_A11.fsa 504.69      NA   1554
#> 20230413_A12.fsa           20230413_A12.fsa 504.29      NA   1272
#> 20230413_B01.fsa           20230413_B01.fsa 563.96      NA    537
#> 20230413_B02.fsa           20230413_B02.fsa 560.83      NA   1031
#> 20230413_B03.fsa           20230413_B03.fsa 561.04      NA    695
#> 20230413_B04.fsa           20230413_B04.fsa 549.17      NA    871
#> 20230413_B05.fsa           20230413_B05.fsa 548.96      NA    913
#> 20230413_B06.fsa           20230413_B06.fsa 543.13      NA    722
#> 20230413_B07.fsa           20230413_B07.fsa 472.28      NA     38
#> 20230413_B08.fsa           20230413_B08.fsa 461.33      NA    665
#> 20230413_B09.fsa           20230413_B09.fsa 463.92      NA    762
#> 20230413_B10.fsa           20230413_B10.fsa 492.43      NA    961
#> 20230413_B11.fsa           20230413_B11.fsa 491.98      NA    879
#> 20230413_B12.fsa           20230413_B12.fsa 482.52      NA   1355
#> 20230413_C01.fsa           20230413_C01.fsa 482.43      NA   1068
#> 20230413_C02.fsa           20230413_C02.fsa 480.10      NA   1164
#> 20230413_C03.fsa           20230413_C03.fsa 480.30      NA   1083
#> 20230413_C04.fsa           20230413_C04.fsa 508.75      NA   1094
#> 20230413_C05.fsa           20230413_C05.fsa 508.54      NA   1123
#> 20230413_C06.fsa           20230413_C06.fsa 509.15      NA   1402
#> 20230413_C07.fsa           20230413_C07.fsa 569.58      NA    718
#> 20230413_C08.fsa           20230413_C08.fsa 569.80      NA    836
#> 20230413_C09.fsa           20230413_C09.fsa 567.50      NA    790
#> 20230413_C10.fsa           20230413_C10.fsa 548.98      NA    835
#> 20230413_C11.fsa           20230413_C11.fsa 546.12      NA    751
#> 20230413_C12.fsa           20230413_C12.fsa 552.04      NA    845
#> 20230413_D01.fsa           20230413_D01.fsa 462.90      NA   1466
#> 20230413_D02.fsa           20230413_D02.fsa 460.65      NA   1375
#> 20230413_D03.fsa           20230413_D03.fsa 460.94      NA   1562
#> 20230413_D04.fsa           20230413_D04.fsa 483.21      NA    676
#> 20230413_D05.fsa           20230413_D05.fsa 480.48      NA    469
#> 20230413_D06.fsa           20230413_D06.fsa 486.52      NA    457
#> 20230413_D07.fsa           20230413_D07.fsa 478.16      NA   1361
#> 20230413_D08.fsa           20230413_D08.fsa 475.33      NA   1382
#> 20230413_D09.fsa           20230413_D09.fsa 475.37      NA   1123
#> 20230413_D10.fsa           20230413_D10.fsa 512.29      NA    692
#> 20230413_D11.fsa           20230413_D11.fsa     NA      NA     NA
#> 20230413_D12.fsa           20230413_D12.fsa 508.37      NA   1230
#> 20230413_E01.fsa           20230413_E01.fsa 563.54      NA    511
#> 20230413_E02.fsa           20230413_E02.fsa 562.65      NA    509
#> 20230413_E03.fsa           20230413_E03.fsa 567.02      NA    595
#> 20230413_E04.fsa           20230413_E04.fsa 549.36      NA    396
#> 20230413_E05.fsa           20230413_E05.fsa 543.62      NA    489
#> 20230413_E06.fsa           20230413_E06.fsa 546.38      NA    515
#> 20230413_E07.fsa           20230413_E07.fsa 458.57      NA   1254
#> 20230413_E08.fsa           20230413_E08.fsa 461.43      NA   1618
#> 20230413_E09.fsa           20230413_E09.fsa 464.23      NA   1494
#> 20230413_E10.fsa           20230413_E10.fsa 478.00      NA    583
#> 20230413_E11.fsa           20230413_E11.fsa 477.82      NA    823
#> 20230413_E12.fsa           20230413_E12.fsa 480.13      NA    989
#> 20230413_F01.fsa           20230413_F01.fsa 476.53      NA    807
#> 20230413_F02.fsa           20230413_F02.fsa 474.28      NA   1001
#> 20230413_F03.fsa           20230413_F03.fsa 477.33      NA    948
#> 20230413_F04.fsa           20230413_F04.fsa 498.71      NA   1075
#> 20230413_F05.fsa           20230413_F05.fsa 502.13      NA    643
#> 20230413_F06.fsa           20230413_F06.fsa 505.53      NA    284
#> 20230413_F07.fsa           20230413_F07.fsa 564.89      NA    651
#> 20230413_F08.fsa           20230413_F08.fsa 564.89      NA    366
#> 20230413_F09.fsa           20230413_F09.fsa 573.83      NA    256
#> 20230413_F10.fsa           20230413_F10.fsa 549.17      NA    536
#> 20230413_F11.fsa           20230413_F11.fsa 542.86      NA    685
#> 20230413_F12.fsa           20230413_F12.fsa 542.65      NA    572
#> 20230413_G01.fsa           20230413_G01.fsa 451.96      NA   1101
#> 20230413_G02.fsa           20230413_G02.fsa 452.61      NA   2877
#> 20230413_G03.fsa           20230413_G03.fsa 452.82      NA    258
#> 20230413_G04.fsa           20230413_G04.fsa 471.85      NA   3452
#> 20230413_G05.fsa           20230413_G05.fsa 471.58      NA   2988
#> 20230413_G06.fsa           20230413_G06.fsa 471.89      NA   2844
#> 20230413_G07.fsa           20230413_G07.fsa 466.33      NA   3363
#> 20230413_G08.fsa           20230413_G08.fsa 466.67      NA   1918
#> 20230413_G09.fsa           20230413_G09.fsa 466.32      NA   3935
#> 20230413_G10.fsa           20230413_G10.fsa 498.92      NA   2957
#> 20230413_G11.fsa           20230413_G11.fsa 498.71      NA    154
#> 20230413_G12.fsa           20230413_G12.fsa 495.11      NA   2950
#> 20230413_H01.fsa           20230413_H01.fsa 545.31      NA   2473
#> 20230413_H02.fsa           20230413_H02.fsa 545.51      NA   1973
#> 20230413_H03.fsa           20230413_H03.fsa 543.00      NA   1896
#> 20230413_H04.fsa           20230413_H04.fsa 524.47      NA   2013
#> 20230413_H05.fsa           20230413_H05.fsa 526.88      NA   1741
#> 20230413_H06.fsa           20230413_H06.fsa 524.38      NA   1340
#> 20230413_H07.fsa           20230413_H07.fsa 414.55      NA   3798
#> 20230413_H08.fsa           20230413_H08.fsa 420.70      NA   2081
#> S-21-211_20220630.fsa S-21-211_20220630.fsa 414.83      NA   3159
#> S-21-212_20220630.fsa S-21-212_20220630.fsa 420.87      NA   2653