Extract ladder summary

Extract a table summarizing the ladder models

Usage

extract_ladder_summary(fragments_trace_list, sort = FALSE)

Arguments

fragments_trace_list

a list of fragments trace objects

sort

A logical statement for if the samples should be ordered by average ladder R-squared.

Value

a dataframe of ladder quality information

Details

The ladder peaks are assigned using a custom algorithm that maximizes the fit of detected ladder peaks and given base-pair sizes. The base pair is assigned using the local Southern method. Basically, for each data point, linear models are made for the lower and upper 3 size standard and the predicted sizes are averaged.

This function summarizes the R-squared values of these individual linear models.

Examples

  fsa_list <- lapply(cell_line_fsa_list, function(x) x$clone())

  find_ladders(fsa_list, show_progress_bar = FALSE)

  extract_ladder_summary(fsa_list, sort = TRUE)
#>                                   unique_id   avg_rsq   min_rsq
#> S-21-211_20220630.fsa S-21-211_20220630.fsa 0.9995315 0.9987599
#> S-21-212_20220630.fsa S-21-212_20220630.fsa 0.9995515 0.9988479
#> 20230413_G01.fsa           20230413_G01.fsa 0.9995583 0.9986045
#> 20230413_F01.fsa           20230413_F01.fsa 0.9995721 0.9989598
#> 20230413_B11.fsa           20230413_B11.fsa 0.9995841 0.9986236
#> 20230413_A12.fsa           20230413_A12.fsa 0.9995856 0.9988091
#> 20230413_H05.fsa           20230413_H05.fsa 0.9995956 0.9989546
#> 20230413_D12.fsa           20230413_D12.fsa 0.9995965 0.9986236
#> 20230413_H07.fsa           20230413_H07.fsa 0.9995995 0.9988479
#> 20230413_D04.fsa           20230413_D04.fsa 0.9996058 0.9986115
#> 20230413_B01.fsa           20230413_B01.fsa 0.9996059 0.9988479
#> 20230413_F12.fsa           20230413_F12.fsa 0.9996078 0.9986236
#> 20230413_E01.fsa           20230413_E01.fsa 0.9996095 0.9988082
#> 20230413_H01.fsa           20230413_H01.fsa 0.9996118 0.9988091
#> 20230413_G02.fsa           20230413_G02.fsa 0.9996129 0.9989598
#> 20230413_B12.fsa           20230413_B12.fsa 0.9996136 0.9989245
#> 20230413_G03.fsa           20230413_G03.fsa 0.9996139 0.9989546
#> 20230413_B02.fsa           20230413_B02.fsa 0.9996151 0.9989546
#> 20230413_G12.fsa           20230413_G12.fsa 0.9996156 0.9988872
#> 20230413_F02.fsa           20230413_F02.fsa 0.9996156 0.9989598
#> 20230413_D11.fsa           20230413_D11.fsa 0.9996160 0.9988872
#> 20230413_E12.fsa           20230413_E12.fsa 0.9996172 0.9989598
#> 20230413_A02.fsa           20230413_A02.fsa 0.9996173 0.9989261
#> 20230413_B10.fsa           20230413_B10.fsa 0.9996189 0.9986115
#> 20230413_H04.fsa           20230413_H04.fsa 0.9996191 0.9987665
#> 20230413_C01.fsa           20230413_C01.fsa 0.9996193 0.9988872
#> 20230413_G11.fsa           20230413_G11.fsa 0.9996193 0.9989261
#> 20230413_A08.fsa           20230413_A08.fsa 0.9996200 0.9986115
#> 20230413_E04.fsa           20230413_E04.fsa 0.9996210 0.9986115
#> 20230413_G09.fsa           20230413_G09.fsa 0.9996218 0.9989546
#> 20230413_H03.fsa           20230413_H03.fsa 0.9996220 0.9988857
#> 20230413_G08.fsa           20230413_G08.fsa 0.9996224 0.9986115
#> 20230413_H02.fsa           20230413_H02.fsa 0.9996232 0.9989245
#> 20230413_D01.fsa           20230413_D01.fsa 0.9996249 0.9988479
#> 20230413_E03.fsa           20230413_E03.fsa 0.9996271 0.9988881
#> 20230413_C02.fsa           20230413_C02.fsa 0.9996303 0.9989546
#> 20230413_A11.fsa           20230413_A11.fsa 0.9996303 0.9989598
#> 20230413_D07.fsa           20230413_D07.fsa 0.9996312 0.9986115
#> 20230413_E02.fsa           20230413_E02.fsa 0.9996313 0.9988479
#> 20230413_B07.fsa           20230413_B07.fsa 0.9996315 0.9986115
#> 20230413_G05.fsa           20230413_G05.fsa 0.9996316 0.9988479
#> 20230413_F10.fsa           20230413_F10.fsa 0.9996318 0.9986115
#> 20230413_A10.fsa           20230413_A10.fsa 0.9996320 0.9989546
#> 20230413_F11.fsa           20230413_F11.fsa 0.9996340 0.9989546
#> 20230413_C07.fsa           20230413_C07.fsa 0.9996343 0.9986115
#> 20230413_A06.fsa           20230413_A06.fsa 0.9996351 0.9989546
#> 20230413_E08.fsa           20230413_E08.fsa 0.9996360 0.9986115
#> 20230413_D03.fsa           20230413_D03.fsa 0.9996374 0.9989546
#> 20230413_G10.fsa           20230413_G10.fsa 0.9996375 0.9989546
#> 20230413_A01.fsa           20230413_A01.fsa 0.9996380 0.9987682
#> 20230413_H08.fsa           20230413_H08.fsa 0.9996397 0.9989546
#> 20230413_F04.fsa           20230413_F04.fsa 0.9996400 0.9989546
#> 20230413_B03.fsa           20230413_B03.fsa 0.9996430 0.9989546
#> 20230413_G07.fsa           20230413_G07.fsa 0.9996454 0.9989546
#> 20230413_F09.fsa           20230413_F09.fsa 0.9996457 0.9988881
#> 20230413_G06.fsa           20230413_G06.fsa 0.9996496 0.9989546
#> 20230413_B04.fsa           20230413_B04.fsa 0.9996499 0.9989546
#> 20230413_A09.fsa           20230413_A09.fsa 0.9996507 0.9989546
#> 20230413_H06.fsa           20230413_H06.fsa 0.9996508 0.9989261
#> 20230413_D02.fsa           20230413_D02.fsa 0.9996510 0.9988479
#> 20230413_F03.fsa           20230413_F03.fsa 0.9996514 0.9989261
#> 20230413_D08.fsa           20230413_D08.fsa 0.9996529 0.9986115
#> 20230413_C06.fsa           20230413_C06.fsa 0.9996536 0.9989278
#> 20230413_E11.fsa           20230413_E11.fsa 0.9996536 0.9989546
#> 20230413_C03.fsa           20230413_C03.fsa 0.9996544 0.9989261
#> 20230413_G04.fsa           20230413_G04.fsa 0.9996552 0.9989261
#> 20230413_F08.fsa           20230413_F08.fsa 0.9996564 0.9989278
#> 20230413_A04.fsa           20230413_A04.fsa 0.9996570 0.9988872
#> 20230413_B09.fsa           20230413_B09.fsa 0.9996575 0.9989546
#> 20230413_E06.fsa           20230413_E06.fsa 0.9996576 0.9989278
#> 20230413_F05.fsa           20230413_F05.fsa 0.9996581 0.9989278
#> 20230413_D05.fsa           20230413_D05.fsa 0.9996581 0.9989546
#> 20230413_C10.fsa           20230413_C10.fsa 0.9996596 0.9989546
#> 20230413_D10.fsa           20230413_D10.fsa 0.9996598 0.9989546
#> 20230413_B05.fsa           20230413_B05.fsa 0.9996600 0.9989546
#> 20230413_A03.fsa           20230413_A03.fsa 0.9996608 0.9989261
#> 20230413_E10.fsa           20230413_E10.fsa 0.9996609 0.9989546
#> 20230413_C12.fsa           20230413_C12.fsa 0.9996609 0.9989623
#> 20230413_C11.fsa           20230413_C11.fsa 0.9996636 0.9990365
#> 20230413_F06.fsa           20230413_F06.fsa 0.9996678 0.9988881
#> 20230413_C05.fsa           20230413_C05.fsa 0.9996684 0.9989546
#> 20230413_C08.fsa           20230413_C08.fsa 0.9996702 0.9989261
#> 20230413_F07.fsa           20230413_F07.fsa 0.9996706 0.9988881
#> 20230413_A05.fsa           20230413_A05.fsa 0.9996732 0.9989646
#> 20230413_E05.fsa           20230413_E05.fsa 0.9996758 0.9989278
#> 20230413_C04.fsa           20230413_C04.fsa 0.9996818 0.9990027
#> 20230413_D06.fsa           20230413_D06.fsa 0.9996841 0.9990027
#> 20230413_A07.fsa           20230413_A07.fsa 0.9996898 0.9990775
#> 20230413_B08.fsa           20230413_B08.fsa 0.9996901 0.9990775
#> 20230413_B06.fsa           20230413_B06.fsa 0.9996961 0.9990775
#> 20230413_E07.fsa           20230413_E07.fsa 0.9996983 0.9989670
#> 20230413_D09.fsa           20230413_D09.fsa 0.9997005 0.9990775
#> 20230413_E09.fsa           20230413_E09.fsa 0.9997041 0.9991142
#> 20230413_C09.fsa           20230413_C09.fsa 0.9997049 0.9990403